async_api_service/pyproject.toml
[tool.poetry]
name = "yandex-async-api"
version = "0.1.0"
description = ""
authors = ["test test <test@test.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
aioredis = "1.3.1"
elasticsearch = {version = "7.9.1", extras = ["async"]}
fastapi = "0.61.1"
pydantic = "1.9.0"
uvicorn = "0.12.2"
asgiref = "3.5.2"
django = "4.1.3"
django-split-settings = "1.2.0"
isort = "5.10.1"
psycopg2-binary = "2.9.5"
python-dotenv = "0.21.0"
sqlparse = "0.4.3"
django-debug-toolbar = "^3.8.1"
django-cors-headers = "^3.13.0"
redis = "^4.4.0"
orjson = "^3.8.5"
uvloop = "^0.17.0"
httptools = "^0.5.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
